Created the abstraction of FhirApplication in core module - engine fails to run in new module
