+++ core/build.gradle
@@ -91,7 +91,10 @@ dependencies {
     implementation deps.room.runtime
     implementation deps.work.runtime
     implementation deps.json_tools.json_patch
-
+    implementation deps.okhttp_logging_interceptor
+    implementation deps.retrofit.core
+    implementation deps.retrofit.gson
+    implementation deps.retrofit.mock
     kapt deps.room.compiler
 
     testImplementation deps.atsl.core
