+++ core/src/main/java/com/google/android/fhir/api/HapiFhirService.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.fhir.reference.api
+package com.google.android.fhir.api
 
 import ca.uhn.fhir.parser.IParser
 import okhttp3.OkHttpClient
