+++ cqlreference/build.gradle
@@ -53,7 +53,7 @@ dependencies {
     testImplementation deps.atsl.core
     implementation deps.constraint_layout
     implementation project(path: ':core')
-
+    implementation deps.work.runtime
     testImplementation deps.junit
 
     androidTestImplementation deps.atsl.ext_junit
