+++ cqlreference/src/main/AndroidManifest.xml
@@ -8,11 +8,8 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.Androidfhir" >
-        <activity
-            android:name="com.google.android.fhir.cqlreference.CqlLoadActivity"
-            android:label="@string/title_cql_load_activity"
-            android:theme="@style/AppTheme.NoActionBar">
+        android:theme="@style/AppTheme">
+        <activity android:name=".CqlLoadActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
