+++ cqlreference/src/main/java/com/google/android/fhir/cqlreference/CqlLoadActivity.kt
@@ -1,23 +1,24 @@
 package com.google.android.fhir.cqlreference
 
 import android.os.AsyncTask
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Button
 import android.widget.EditText
 import android.widget.TextView
-import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.ViewModelProvider
 import ca.uhn.fhir.context.FhirContext
 import com.google.android.fhir.FhirEngine
-import com.google.android.fhir.reference.FhirApplication.Companion.fhirEngine
+import com.google.android.fhir.cqlreference.FhirApplication.Companion.fhirEngine
 import com.google.android.material.snackbar.Snackbar
+import org.hl7.fhir.r4.model.Resource
+import org.opencds.cqf.cql.execution.EvaluationResult
 import java.io.BufferedReader
 import java.io.IOException
 import java.io.InputStream
 import java.io.InputStreamReader
 import java.net.URL
-import org.hl7.fhir.r4.model.Resource
-import org.opencds.cqf.cql.execution.EvaluationResult
+
 
 class CqlLoadActivity : AppCompatActivity() {
   lateinit var fhirEngine: FhirEngine
