+++ cqlreference/src/main/java/com/google/android/fhir/cqlreference/FhirApplication.kt
@@ -1,20 +1,4 @@
-/*
- * Copyright 2020 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.google.android.fhir.reference
+package com.google.android.fhir.cqlreference
 
 import android.content.Context
 import androidx.work.Constraints
@@ -36,7 +20,6 @@ import org.hl7.fhir.r4.model.ResourceType
 
 class FhirApplication : CoreApplication() {
 
-  // only initiate the FhirEngine when used for the first time, not when the app is created
   override val fhirEngine: FhirEngine by lazy { constructFhirEngine() }
 
   override fun constructFhirEngine(): FhirEngine {
