+++ reference/src/main/AndroidManifest.xml
@@ -1,25 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
-<!--
- Copyright 2020 Google LLC
-
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
--->
-
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.google.android.fhir.reference">
+
     <uses-permission android:name="android.permission.INTERNET" />
+
     <application
-        android:name="com.google.android.fhir.reference.FhirApplication"
+        android:name=".FhirApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -27,9 +13,9 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-                android:name="com.google.android.fhir.reference.PatientListActivity"
-                android:label="@string/title_patient_list"
-                android:theme="@style/AppTheme.NoActionBar">
+            android:name=".PatientListActivity"
+            android:label="@string/title_patient_list"
+            android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -37,13 +23,14 @@
             </intent-filter>
         </activity>
         <activity
-            android:name="com.google.android.fhir.reference.PatientDetailActivity"
+            android:name=".PatientDetailActivity"
             android:label="@string/title_patient_detail"
-            android:parentActivityName="com.google.android.fhir.reference.PatientListActivity"
+            android:parentActivityName=".PatientListActivity"
             android:theme="@style/AppTheme.NoActionBar">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value="com.google.android.fhir.reference.PatientListActivity" />
         </activity>
     </application>
+
 </manifest>
