+++ reference/src/main/res/menu/list_options_menu.xml
@@ -24,10 +24,4 @@
         android:icon="@android:drawable/stat_notify_sync"
         android:title="@string/menu_sync_resources"
         />
-    <item
-        android:id="@+id/load_resource"
-        app:showAsAction="withText"
-        android:icon="@android:drawable/ic_menu_add"
-        android:title="@string/menu_load_resource"
-        />
 </menu>
